pyproject.toml
[tool.poetry]
name = "pixiv-loli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
peewee = "^3.14.8"
fastapi = {extras = ["all"], version = "^0.70.0"}
apscheduler = "^3.8.1"
func-timeout = "^4.3.5"
pixivpy = "^3.6.1"
pillow = "^9.2.0"
tencentcloud-sdk-python = "^3.0.598"
slowapi = "^0.1.5"
fastapi-limiter = "^0.1.4"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true